Fix consensus strategy monkeypatch compatibility
